add relay style post pagination

+++ server/src/services/base.service.ts
@@ -1,7 +1,7 @@
-import { DataSource, ObjectLiteral } from "typeorm";
+import { DataSource, ObjectLiteral, Repository } from "typeorm";
 
-export abstract class BaseService<T> {
-  protected readonly repo: ObjectLiteral;
+export abstract class BaseService<T extends ObjectLiteral> {
+  protected readonly repo: Repository<T>;
 
   constructor(
     protected readonly AppDataSource: DataSource,
@@ -10,16 +10,24 @@ export abstract class BaseService<T> {
     this.repo = AppDataSource.getRepository(entity);
   }
 
+  get repository() {
+    return this.repo;
+  }
+
   async findById(id: number): Promise<T | null> {
-    return this.repo.findOneBy({ id });
+    return this.repo.findOneBy({ id } as any);
   }
 
   async find(): Promise<T[]> {
     return this.repo.find();
   }
 
-  async remove(id: number): Promise<number> {
-    const entity = this.repo.findById(id);
+  async remove(id: number): Promise<T | null> {
+    const entity = await this.findById(id);
+
+    if (!entity) {
+      return null;
+    }
 
     return this.repo.remove(entity);
   }
